Fix seconds output in common lisp solution
get-internal-real-time returns the time in "internal time units" which are lisp
implementation dependent, not milliseconds. On my Debian GNU/Linux system with sbcl
2.2.3, it was not milliseconds and in fact was causing the output to
be 900 seconds vs 0.9 seconds (a factor of 1000). Interestingly, the docker build
doesn't have this same issue.

Using internal-time-units-per-second should fix this in all cases since it is
defined by the standard.

http://www.lispworks.com/documentation/lw51/CLHS/Body/f_get_in.htm
http://www.lispworks.com/documentation/lw51/CLHS/Body/v_intern.htm

Fix tested in docker, I have not tested the entire script
for the project.